core: Rename source files and iterator variables
