+++ src/core/magnetostatics/CMakeLists.txt
@@ -20,8 +20,8 @@ target_sources(
   espresso_core
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/dipoles.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/barnes_hut_gpu.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/dds.cpp
-          ${CMAKE_CURRENT_SOURCE_DIR}/dds_gpu.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/dipolar_direct_sum.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/dipolar_direct_sum_gpu.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/dlc.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/dp3m.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/scafacos_impl.cpp)
