+++ src/core/magnetostatics/dipolar_direct_sum.hpp
