+++ src/core/magnetostatics/dipolar_direct_sum_gpu_cuda.cuh
